server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via
    #   django
    #   uvicorn
billiard==3.6.4.0
    # via celery
black==22.3.0
    # via -r requirements.in
boto3==1.24.19
    # via -r requirements.in
botocore==1.27.20
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via -r requirements.in
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
django==4.0.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.2.4
    # via celery
mypy-extensions==0.4.3
    # via black
mysqlclient==2.1.0
    # via -r requirements.in
numpy==1.23.0
    # via
    #   -r requirements.in
    #   opencv-python
opencv-python==4.6.0.66
    # via -r requirements.in
packaging==21.3
    # via pytesseract
pathspec==0.9.0
    # via black
pdf2image==1.16.0
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pillow==9.1.1
    # via
    #   pdf2image
    #   pytesseract
pip-tools==6.6.2
    # via -r requirements.in
platformdirs==2.5.2
    # via black
prompt-toolkit==3.0.29
    # via click-repl
pyparsing==3.0.9
    # via packaging
pytesseract==0.3.9
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
pytz==2022.1
    # via
    #   celery
    #   djangorestframework
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.2
    # via django
tomli==2.0.1
    # via
    #   black
    #   pep517
typing-extensions==4.2.0
    # via black
urllib3==1.26.9
    # via botocore
uvicorn==0.17.6
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
